+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Change log
 
+## [Unreleased]
+### Changed
+- NSR quay info mapped into the Calls table.
+
 ## [0.1.0] – 2024-04-18
 ### Added
 - Implemented sink API for fetching and deleting raw data